requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
streamlit
requests
python-dotenv